Changed apdex rate to decimal. At the ES side it represent  as double

+++ src/Memores.Metrics.Wcf/Model/Reports/ApdexReport.cs
@@ -2,6 +2,6 @@
 
 namespace Memores.Metrics.Wcf.Model.Reports {
     public class ApdexReport: MetricsReportBase {
-        public long Apdex { get; set; }
+        public decimal Apdex { get; set; }
     }
 }
